port number

+++ src/hub.c
@@ -213,7 +213,7 @@ if (thislock.lock != NULL)
 aa.copy.portnumber = SHORT_CFENGINEPORT;
 
 CfOut(cf_inform,"","...........................................................................\n");
-CfOut(cf_inform,""," * Hailing %s : %u\n",peer,aa.copy.portnumber);
+CfOut(cf_inform,""," * Hailing %s : %u\n",peer,(int)aa.copy.portnumber);
 CfOut(cf_inform,"","...........................................................................\n");
 
 /* Check trust interaction*/
